Allow the fulfillment of a continuable to say whether it was successful or not.
Before this was entirely dependent on whether or not the the fulfilled
value was instanceof Error.  Now, when you originally fulfill a
continuable you can pass true or false as the second parameter to
indicate if it was successful or not.

At this point if you are chaining the continuables then the only way
from a link in the chain to specify that it wasn't successful is to
return an object that is instanceof Error.  This is because you cannot
return more than one thing from a function.

This change is a follow up to the previous commit, and makes it easier
to work with Promises which when they emitError, they don't emit an object
that is instanceof Error.  Now, the success parameter passed to the
callback will be set appropriately.

It is still highly encouraged that if you are going to be fulfilling a
continuable with something that should have a false success parameter
use an object that is instanceof Error.  This is just good practice.